refactor: put all validation files into a separate folder
put all validation files into a separate folder

MODULE-1-PART-3-gallery
